# docs/swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/rest/v1/sheets": {
      "get": {
        "description": "This will show all sheets by default.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "sheets"
        ],
        "summary": "Get all sheets",
        "operationId": "getAllSheets",
        "responses": {
          "200": {
            "$ref": "#/responses/SheetResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": []
          }
        ]
      },
      "post": {
        "description": "This will create a new sheet.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "sheets"
        ],
        "summary": "Create sheets",
        "operationId": "createSheet",
        "parameters": [
          {
            "description": "The request body for creating a sheet",
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateSheetRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/SheetResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/responses/ErrorResponse"
          },
          "413": {
            "$ref": "#/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": []
          }
        ]
      }
    },
    "/rest/v1/sheets/{id}": {
      "get": {
        "description": "This will get the sheet with the given id.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "sheets"
        ],
        "summary": "Get sheets",
        "operationId": "getSheet",
        "parameters": [
          {
            "type": "integer",
            "format": "int32",
            "description": "id of the sheet to get",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SheetResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": []
          }
        ]
      },
      "put": {
        "description": "This will update a sheet by its given id.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "sheets"
        ],
        "summary": "Update sheets",
        "operationId": "updateSheetById",
        "parameters": [
          {
            "description": "The request body for updating a sheet",
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateSheetByIdRequest"
            }
          },
          {
            "type": "integer",
            "format": "int32",
            "description": "id of the sheet to update",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/SheetResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          },
          "409": {
            "$ref": "#/responses/ErrorResponse"
          },
          "413": {
            "$ref": "#/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": []
          }
        ]
      },
      "delete": {
        "description": "This will delete a sheet by its given id.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "sheets"
        ],
        "summary": "Delete sheets",
        "operationId": "deleteSheet",
        "parameters": [
          {
            "type": "integer",
            "format": "int32",
            "description": "id of the sheet to delete",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ItemDeletedResponse represents the response when an item is deleted",
            "schema": {
              "$ref": "#/definitions/ItemDeletedResponse"
            }
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": []
          }
        ]
      }
    },
    "/rest/v1/sheets/{id}/shots": {
      "get": {
        "description": "This will return every shot for the sheet with the given id, as a JSON\narray with the same shape as GET /rest/v1/shots. Returns an empty array\nfor an existing sheet without shots.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "sheets"
        ],
        "summary": "Get shots by sheet id",
        "operationId": "getShotsBySheetId",
        "parameters": [
          {
            "type": "integer",
            "format": "int32",
            "description": "id of the sheet whose shots to get",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ShotResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          },
          "404": {
            "$ref": "#/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "oauth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "CreateSheetRequest": {
      "description": "CreateSheetRequest represents the request body for creating a sheet",
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/lescactus/espressoapi-go/internal/controllers/rest"
    },
...
    "ItemDeletedResponse": {
      "description": "ItemDeletedResponse represents the response when an item is deleted",
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Id"
        },
        "msg": {
          "type": "string",
          "x-go-name": "Msg"
        }
      },
      "x-go-package": "github.com/lescactus/espressoapi-go/internal/controllers/rest"
    },
...
    "UpdateSheetByIdRequest": {
      "description": "UpdateSheetByIdRequest represents the request body for updating a sheet\nwith the given id",
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/lescactus/espressoapi-go/internal/controllers/rest"
    },
...
  },
  "responses": {
...
    "ErrorResponse": {
      "description": "ErrorResponse represents the json response\nfor http errors.\nIt contains a message describing the error",
      "headers": {
        "msg": {
          "type": "string"
        }
      }
    },
...
    "SheetResponse": {
      "description": "SheetResponse represents a sheet for this application\n\nA sheet is a collection of shots. It's used to group shots together\nin a logical way.",
      "headers": {
        "created_at": {
          "type": "string",
          "format": "date-time",
          "description": "The creation date of the sheet"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "The id for the sheet"
        },
        "name": {
          "type": "string",
          "description": "The name for the sheet"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "description": "The last update date of the sheet"
        }
      }
    },
    "ShotResponse": {
      "description": "ShotResponse represents an espresso shot for this application\n\nAn espresso shot is made from coffee beans, ground at a specific setting,\nwith a specific quantity of coffee in and out.\nIt also has a specific shot time and water temperature.\n\nThe result of a shot can be rated and compared to the previous shot.\nIt can also be too bitter or too sour.",
      "headers": {
        "additional_notes": {
          "type": "string"
        },
        "beans": {},
        "comparison_with_previous_result": {
          "type": "integer",
          "format": "uint8"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "grind_setting": {
          "type": "integer",
          "format": "int64"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "is_too_bitter": {
          "type": "boolean"
        },
        "is_too_sour": {
          "type": "boolean"
        },
        "quantity_in": {
          "type": "number",
          "format": "double"
        },
        "quantity_out": {
          "type": "number",
          "format": "double"
        },
        "rating": {
          "type": "number",
          "format": "double"
        },
        "sheet": {},
        "shot_time": {
          "type": "integer",
          "format": "int64"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        },
        "water_temperature": {
          "type": "number",
          "format": "double"
        }
      }
    }
  }
}